+++ src/lang-xquery/main/uk/co/reecedunn/intellij/plugin/xquery/psi/impl/xquery/XQueryValidateExprPsiImpl.kt
@@ -24,7 +24,13 @@ import uk.co.reecedunn.intellij.plugin.xquery.ast.xquery.XQueryValidateExpr
 import uk.co.reecedunn.intellij.plugin.xpath.lexer.XPathTokenType
 import uk.co.reecedunn.intellij.plugin.xquery.lexer.XQueryTokenType
 
-private val VALIDATE_BY_TYPENAME = TokenSet.create(XPathTokenType.K_AS, XQueryTokenType.K_TYPE)
+private val VALIDATE_BY = TokenSet.create(
+    XQueryTokenType.K_STRICT,
+    XQueryTokenType.K_LAX,
+    XQueryTokenType.K_FULL,
+    XPathTokenType.K_AS,
+    XQueryTokenType.K_TYPE
+)
 
 private val XQUERY10: List<Version> = listOf()
 private val XQUERY30: List<Version> = listOf(XQuerySpec.REC_3_0_20140408, MarkLogic.VERSION_6_0)
@@ -35,13 +41,13 @@ class XQueryValidateExprPsiImpl(node: ASTNode) : ASTWrapperPsiElement(node), XQu
         get(): List<Version> {
             val element = conformanceElement
             if (element !== firstChild) {
-                if (element.node.elementType === XQueryTokenType.K_TYPE) {
-                    return XQUERY30
+                return when (element.node.elementType) {
+                    XPathTokenType.K_AS, XQueryTokenType.K_FULL -> MARKLOGIC60
+                    else -> XQUERY30
                 }
-                return MARKLOGIC60
             }
             return XQUERY10
         }
 
-    override val conformanceElement get(): PsiElement = findChildByType(VALIDATE_BY_TYPENAME) ?: firstChild
+    override val conformanceElement get(): PsiElement = findChildByType(VALIDATE_BY) ?: firstChild
 }