+++ CHANGELOG.md
@@ -61,6 +61,8 @@ BaseX:
 MarkLogic:
 
 1.  Add definitions for the MarkLogic 10.0-1 built-in functions.
+1.  Fix MarkLogic 6.0 `binary()` used as a `NodeTest`.
+1.  Support the MarkLogic 6.0 `validate full` syntax extension.
 
 XPath and XQuery:
 
@@ -72,7 +74,6 @@ XPath and XQuery:
     non-delimiting terminal symbols.
 1.  Report an error if an unknown axis name is followed by the `::` axis indicator.
 1.  Full Text 1.0 and 3.0 extensions support in the XPath lexer and parser.
-1.  Fix MarkLogic 6.0 `binary()` used as a `NodeTest`.
 
 Inspections: